requirements.txt
discord.py>=2.3.2
requests>=2.31.0
pandas>=2.0.0
scikit-learn>=1.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
joblib>=1.3.0
Flask>=3.0.0
google-generativeai>=0.8.0
